requirements.txt
absl-py
astor
astunparse
cached-property
cachetools
certifi
charset-normalizer
cloudpickle
cycler
dataclasses
decorator
dill
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
googleapis-common-protos
h5py
idna
importlib-metadata
importlib-resources
keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
networkx
oauthlib
opt-einsum
Pillow
promise
protobuf
psutil
pyasn1
pyasn1-modules
pyparsing
python-dateutil
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-addons
tensorflow-datasets
tensorflow-estimator
tensorflow-graphics
tensorflow-metadata
tensorflow-probability
termcolor
tqdm
trimesh
typeguard
typing-extensions
wrapt
zipp

